+++ cms/middleware.py
@@ -13,15 +13,6 @@ if 'cms.middleware.VersionMiddleware' in settings.MIDDLEWARE:
     from .apps.pages.middleware import VersionMiddleware
 
 
-def get_client_ip(request):
-    x_forwarded_for = request.META.get('HTTP_X_FORWARDED_FOR')
-    if x_forwarded_for:
-        ip = x_forwarded_for.split(',')[0]
-    else:
-        ip = request.META.get('REMOTE_ADDR')
-    return ip
-
-
 class PublicationMiddleware:
 
     '''Middleware that enables the preview mode for admin users.'''
